# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gerenciador de Produtos - API",
        "description": "API REST completa para gerenciamento de produtos, autenticação e usuários. Implementada em Laravel 10 com Sanctum para autenticação stateless.",
        "contact": {
            "email": "admin@gerenciador-produtos.com"
        },
        "license": {
            "name": "MIT License"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar produtos",
                "description": "Retorna lista paginada de produtos com suporte a filtros opcionais",
                "operationId": "b44f47e2fa424c65d2db9aba57590885",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nome do produto",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "preco_min",
                        "in": "query",
                        "description": "Preço mínimo",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "preco_max",
                        "in": "query",
                        "description": "Preço máximo",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "estoque_min",
                        "in": "query",
                        "description": "Estoque mínimo",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "estoque_max",
                        "in": "query",
                        "description": "Estoque máximo",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produtos listados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Criar novo produto",
                "description": "Cria um novo produto na base de dados",
                "operationId": "a02ae3401c5ed84c9d6e582ee8e968bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "preco",
                                    "quantidade_estoque"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Teclado Mecânico"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 450
                                    },
                                    "quantidade_estoque": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Teclado RGB com switches mecânicos",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/produtos/{produto}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Obter produto por ID",
                "description": "Retorna os dados detalhados de um produto específico",
                "operationId": "154beb0cd18d5cf056423c349c73a1d2",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar produto",
                "description": "Atualiza os dados de um produto existente",
                "operationId": "225048246a0ca0c827fc8b8b89dab43e",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Teclado Mecânico RGB"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "quantidade_estoque": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Teclado RGB com switches mecânicos premium",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Deletar produto",
                "description": "Deleta um produto (soft delete - pode ser restaurado)",
                "operationId": "ada40905f121246cd27f06dfaa907b27",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/produtos/trashed": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar produtos deletados",
                "description": "Retorna lista paginada de produtos deletados (soft delete)",
                "operationId": "9b75f6e1fb21fa0ce8acbe3e84ed6a69",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produtos deletados listados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/produtos/{produto}/restore": {
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Restaurar produto",
                "description": "Restaura um produto que foi deletado (soft delete)",
                "operationId": "16c723c97210182a3a4b6b6c361895a3",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "path",
                        "description": "ID do produto deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto restaurado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Bearer (obtido no login)",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}